+++ src/routes/main.js
@@ -1,4 +1,5 @@
 import { setGetForm, setObjectView } from "./get/get.js";
+import { setPostForm, setPostView } from "./post/post.js";
 
 let isResumeDeleted = false;
 let isStylingDeleted = false;
@@ -69,29 +70,29 @@ function resetForm(element) {
   }
 }
 
-function setPostForm() {
-  // resetForm(document.getElementById("demo"));
-  document.getElementById("demo").innerHTML = "...personal-site/resource?type=";
-  var form = document.createElement("form");
-  form.id = "send-object";
-  select = document.createElement("select");
-  select.id = "object-choice";
-  form.appendChild(select);
+// function setPostForm() {
+//   // resetForm(document.getElementById("demo"));
+//   document.getElementById("demo").innerHTML = "...personal-site/resource?type=";
+//   var form = document.createElement("form");
+//   form.id = "send-object";
+//   select = document.createElement("select");
+//   select.id = "object-choice";
+//   form.appendChild(select);
 
-  optionDefault = document.createElement("option");
-  optionDefault.value = "";
-  optionDefault.hidden = true;
-  optionDefault.disabled = true;
-  optionDefault.selected = true;
-  select.appendChild(optionDefault);
+//   optionDefault = document.createElement("option");
+//   optionDefault.value = "";
+//   optionDefault.hidden = true;
+//   optionDefault.disabled = true;
+//   optionDefault.selected = true;
+//   select.appendChild(optionDefault);
 
-  option1 = document.createElement("option");
-  option1.value = "contact";
-  option1.innerHTML = "Contact-Me";
-  select.appendChild(option1);
+//   option1 = document.createElement("option");
+//   option1.value = "contact";
+//   option1.innerHTML = "Contact-Me";
+//   select.appendChild(option1);
 
-  document.getElementById("demo").after(form);
-}
+//   document.getElementById("demo").after(form);
+// }
 
 function setPatchForm() {}
 
@@ -131,17 +132,17 @@ function setDeleteForm() {
   document.getElementById("demo").after(form);
 }
 
-function setPostView(object) {
-  document.getElementById("render-object").innerHTML = "";
-  switch (object) {
-    case "contact":
-      emailObject = document.createElement("a");
-      emailObject.href = "mailto:aliladha2000@gmail.com";
-      emailObject.innerHTML = "Email";
-      document.getElementById("render-object").appendChild(emailObject);
-      break;
-  }
-}
+// function setPostView(object) {
+//   document.getElementById("render-object").innerHTML = "";
+//   switch (object) {
+//     case "contact":
+//       emailObject = document.createElement("a");
+//       emailObject.href = "mailto:aliladha2000@gmail.com";
+//       emailObject.innerHTML = "Email";
+//       document.getElementById("render-object").appendChild(emailObject);
+//       break;
+//   }
+// }
 
 function setDeleteDescription(object) {
   document.getElementById("render-object").innerHTML = "";
